Add files via upload
This is the new and improved script for tracking theoretical jumps by drawing lines!
Summary: 
Attach this script to the player object or any child object of the player and supply it with a line prefab of your choice. Then, when the game is running in play mode it will make a series of lines showing the path of the player if it were to move at full speed to the right and jump as many times as possible as soon as they were able.
This script accounts for any changes made to the player controller script affecting:
-Jump Force
-Player Speed
-Number of jumps
-Time between jumps
-Gravity Modifier (Note: since gravity modifier is invoked at start, the linemaker will reflect changes made to the gravity modifier after play starts even if changing the modifier at that point doesn't do anything)

You do not need the previous LineGuide script to operate this script but for full functionality you will want to give your line prefab the GuideLine tag

Warnings:
-Does NOT currently take object mass or drag into account. Script assumes an object mass of 1 and drag of 0. This is due to not interfacing with the rigidbody components much at all and is only really a problem due to forcetype differences.
-Does NOT innately track any new movement types (i.e. jumps using different forcetypes, gliding, dashing, etc.) such functionalities would have to be added later in the program's current state.

I hope y'all like this and find it somewhat useful. I definitely learned a lot while making it but I do have to admit the drawbacks and time investment were less than ideal.